.vscode/requirements.txt
fastapi==0.74.1
passlib==1.7.4
pydantic==1.9.0
PyJWT==2.3.0
pymongo[srv]
PyYAML==6.0
starlette==0.17.1
uvicorn==0.11.8
fastapi-mail
bcrypt